추가 패치, 머지 위한 처리

+++ src/main/java/com/team5/WalkingWithWorld/comments/entity/Comments.java
@@ -16,7 +16,7 @@ import java.time.LocalDateTime;
 public class Comments extends AuditingFields {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
     @ManyToOne(optional = false)
     @JoinColumn(name = "users_id")
@@ -29,7 +29,7 @@ public class Comments extends AuditingFields {
     private String content;
 
 
-    public static Comments of (int id,
+    public static Comments of (Long id,
                                Users users,
                                WalkingPaths walkingPaths,
                                String content
