+++ src/main/java/com/team5/WalkingWithWorld/comments/repository/CommentsRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 
 import java.util.List;
 @Repository
-public interface CommentsRepository extends JpaRepository<Comments, Integer> {
+public interface CommentsRepository extends JpaRepository<Comments, Long> {
     List<Comments> findTop5ByWalkingPathsIdOrderByIdDesc(int id);
 }
